Add files via upload
Created the main class and randomGenerate class that will handle all the generating of the password give. The main method is mostly for inputs and outputs. Tested the random generation of numbers specifically but it seems like it runs but doesn't output anything. So once I find a way for each letters, special letters, and numbers to generate and able to print output then I will add all the strings up together then randomize each again.